Bug fix -Only Image owner should be allow to edit/delete image

+++ ImageHoster_assignment_stub/ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,29 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
-        Image image = imageService.getImage(imageId);
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        Image image = imageService.getImage(imageId);
+        User user=(User)session.getAttribute("loggeduser");
+
+        String error = "Only the owner of the image can edit the image";
+        if(user.getUsername().equals(image.getUser().getUsername())){
+
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+
+        else{
+
+            model.addAttribute("editError", error);
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            return "images/image";
+        }
+
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -132,7 +148,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" +imageId+"/"+updatedImage.getTitle();
     }
 
 
@@ -140,9 +156,23 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session,Model model) {
+
+        User user=(User)session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        String error = "Only the owner of the image can delete the imag";
+
+        if(image.getUser().getUsername().equals(user.getUsername())){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else{
+            model.addAttribute("deleteError",error);
+            model.addAttribute("image",image);
+            model.addAttribute("tags", image.getTags());
+            return "images/image";
+        }
+
     }
